Using Power-of-Two textures with GLES 2.0 backend

+++ src/File.cpp
@@ -198,7 +198,7 @@ uint64_t File::Seek( int64_t ofs, DIR dir )
 			mOffset = ofs;
 		} else if ( dir == CURR && (int64_t)mOffset + ofs >= 0 && (int64_t)mOffset + ofs < (int64_t)mBufferSize ) {
 			mOffset = mOffset + ofs;
-		} else if ( dir == END && (int64_t)mBufferSize + ofs >= 0 && mBufferSize + ofs < (int64_t)mBufferSize ) {
+		} else if ( dir == END && (int64_t)mBufferSize + ofs >= 0 && (int64_t)mBufferSize + ofs < (int64_t)mBufferSize ) {
 			mOffset = mBufferSize + ofs;
 		}
 		return mOffset;
